keep a running memory summary which will replace fully decayed memories with the summary in its usage
after fully decayed memories ( > 20 summarizations @ 2 weeks of non recall each) then we delete memories which would force aida to recall the summary. The quality of the summary will be better because it summarizes on the spot in topic form and only against other important memories to avoid nonsense being added to generalize memories.